Finalizando CRUD

+++ src/main/java/med/voll/api/medico/DadosListagemMedico.java
@@ -1,6 +1,6 @@
 package med.voll.api.medico;
 
-public record DadosListagemMedico(Long id, String nome, String email, String crm, Especialidade especialidade) {
+public record DadosListagemMedico(Integer id, String nome, String email, String crm, Especialidade especialidade) {
 
     public DadosListagemMedico(Medico medico) {
         this(medico.getId(), medico.getNome(), medico.getEmail(), medico.getCrm(), medico.getEspecialidade());
